Create FilesToTape_TodaysFull.ps1
This ps1 file simplifies tape backups with Veeam Backup & Replication when only the most recent full backup is required, and the original repository cannot be emptied to synchronize the Backup job and the Tape Backup job.

This script works on date matching. If today's date is different from the date in the name of the backup file you want to move, the file will not be moved or backed up.